requirements.txt
pymupdf==1.22.5
python-docx==1.1.0
openpyxl==3.1.2
pandas==1.5.3
python-pptx==0.6.21

# pip install -r requirements.txt